Finish PHP

+++ CMS_TEMPLATE/admin/functions_posts.php
@@ -80,7 +80,7 @@ function ViewAllPosts(){
             echo "<td>$post_comment_count</td>";
             echo "<td>$post_date</td>";
             echo "<td><a href='./posts.php?source=edit_post&edit={$post_id}'>Edit</a></td>";
-            echo "<td><a href='./posts.php?delete={$post_id}'>Delete</a></td>";         
+            echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete'); \" href='./posts.php?delete={$post_id}'>Delete</a></td>";         
             echo "</tr>";
         }
